+++ WindowsApplication/scripts/ScriptTemplate.h
@@ -17,6 +17,6 @@ public:
 	~ScriptTemplate();
 
 	void Start();
-	void Update(int deltaTime);
+	void Update(float deltaTime);
 };
